fix(ue): Move actors to NR UE UE_thread stack.
This prevents a segfault in case nf is used from sync actor during teardown.
