+++ executables/nr-uesoftmodem.c
@@ -63,7 +63,6 @@ unsigned short config_frames[4] = {2,9,11,13};
 #include "executables/thread-common.h"
 
 #include "nr_nas_msg.h"
-#include "actor.h"
 
 THREAD_STRUCT thread_struct;
 nrUE_params_t nrUE_params = {0};
@@ -139,67 +138,6 @@ static void get_options(configmodule_interface_t *cfg)
   config_get(cfg, cmdline_params, numparams, NULL);
   AssertFatal(nrUE_params.extra_pdu_id == -1,
               "Add additional PDU sessions in uicc.pdu_sessions array instead\n");
-}
-
-/* Parse --actor-affinity into cores[]. Returns number of cores. Empty/NULL -> 0. */
-static int parse_actor_affinity(const char *params, int *cores, int max_cores)
-{
-  if (params == NULL || params[0] == '\0')
-    return 0;
-
-  char *cpy = strdup(params);
-  AssertFatal(cpy != NULL, "Memory exhausted\n");
-  int n = 0;
-  char *saveptr = NULL;
-  for (char *tok = strtok_r(cpy, ",", &saveptr); tok != NULL; tok = strtok_r(NULL, ",", &saveptr)) {
-    AssertFatal(n < max_cores, "actor-affinity has more than %d entries\n", max_cores);
-    cores[n++] = atoi(tok);
-  }
-  free(cpy);
-  return n;
-}
-
-/* Assign one core from the shared pool to each DL actor, then each UL actor.
- * If a spare remains, pin SYNC too. Leave everyone at -1 when actor-affinity is unset. */
-static void init_ue_actors(PHY_VARS_NR_UE *UE)
-{
-  const nrUE_params_t *p = get_nrUE_params();
-  const int need = p->num_dl_actors + p->num_ul_actors;
-  int cores[64];
-  const int have = parse_actor_affinity(p->actor_affinity, cores, sizeofArray(cores));
-  AssertFatal(have == 0 || have >= need,
-              "actor-affinity has %d entries but needs at least num-dl-actors (%d) + num-ul-actors (%d) = %d\n",
-              have,
-              p->num_dl_actors,
-              p->num_ul_actors,
-              need);
-
-  int idx = 0;
-  const int sync_core = (have > need) ? cores[need] : -1;
-  init_actor(&UE->sync_actor, "SYNC_", sync_core);
-
-  if (p->num_dl_actors > 0) {
-    UE->dl_actors = calloc_or_fail(p->num_dl_actors, sizeof(*UE->dl_actors));
-    for (int i = 0; i < p->num_dl_actors; i++) {
-      const int core = (have > 0) ? cores[idx++] : -1;
-      init_actor(&UE->dl_actors[i], "DL_", core);
-    }
-  }
-  if (p->num_ul_actors > 0) {
-    UE->ul_actors = calloc_or_fail(p->num_ul_actors, sizeof(*UE->ul_actors));
-    for (int i = 0; i < p->num_ul_actors; i++) {
-      const int core = (have > 0) ? cores[idx++] : -1;
-      init_actor(&UE->ul_actors[i], "UL_", core);
-    }
-  }
-  if (have > 0)
-    LOG_I(PHY,
-          "Pinned actors from actor-affinity=%s (DL %d, UL %d, SYNC %d)%s\n",
-          p->actor_affinity,
-          p->num_dl_actors,
-          p->num_ul_actors,
-          sync_core,
-          have > need + 1 ? " (extra cores unused)" : "");
 }
 
 // set PHY vars from command line
@@ -436,7 +374,6 @@ int main(int argc, char **argv)
       mac->dl_frequency = cell.rf_frequency;
 
       UE_CC->sl_mode = get_softmodem_params()->sl_mode;
-      init_ue_actors(UE_CC);
       init_nr_ue_vars(UE_CC, inst);
 
       if (UE_CC->sl_mode) {
@@ -506,12 +443,6 @@ int main(int argc, char **argv)
     for (int CC_id = 0; CC_id < MAX_NUM_CCs; CC_id++) {
       PHY_VARS_NR_UE *phy_vars = nrPHY_vars_UE_g[0][CC_id];
       if (phy_vars) {
-        for (int i = 0; i < get_nrUE_params()->num_ul_actors; i++) {
-          shutdown_actor(&phy_vars->ul_actors[i]);
-        }
-        for (int i = 0; i < get_nrUE_params()->num_dl_actors; i++) {
-          shutdown_actor(&phy_vars->dl_actors[i]);
-        }
         int ret = pthread_join(phy_vars->main_thread, NULL);
         AssertFatal(ret == 0, "pthread_join error %d, errno %d (%s)\n", ret, errno, strerror(errno));
         if (!IS_SOFTMODEM_NOSTATS) {
