+++ openair1/PHY/defs_nr_UE.h
@@ -442,9 +442,6 @@ typedef struct PHY_VARS_NR_UE_s {
   // Sidelink parameters
   sl_nr_sidelink_mode_t sl_mode;
   sl_nr_ue_phy_params_t SL_UE_PHY_PARAMS;
-  Actor_t sync_actor;
-  Actor_t *dl_actors;
-  Actor_t *ul_actors;
   pthread_t main_thread;
   pthread_t stat_thread;
   // Per-DL-actor pre-allocated PDSCH scratch buffers (one set per actor to avoid races)
@@ -598,6 +595,7 @@ typedef struct nr_rxtx_thread_data_s {
   nr_phy_data_t phy_data;
   dynamic_barrier_t* next_barrier;
   uint64_t absolute_deadline_us;
+  Actor_t *ul_actors;
 } nr_rxtx_thread_data_t;
 
 typedef struct LDPCDecode_ue_s {
